+++ src/components/App/App.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Switch, Route } from "react-router-dom";
-import Home from './Home';
-import Header from './Header';
-import ProfileVisualization from './ProfileVisualization';
-import ErrorPage from './ErrorPage'
+import Home from '../Home/Home';
+import Header from '../Header/Header';
+import ProfileVisualization from '../ProfileVisualization/ProfileVisualization';
+import ErrorPage from '../ErrorPage/ErrorPage'
 import './App.css';
 
 const App = () => {
