+++ src/components/Home/Home.css
@@ -50,7 +50,7 @@ main {
   width: 90vw;
   max-width: 500px;
   min-height: 35vh;
-  background-image: url('./shoveling_earth.png');
+  background-image: url('../assets/shoveling_earth.png');
   background-position: center;
   background-size: 100%;
   background-repeat: no-repeat;
