+++ src/components/Home/Home.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Link } from "react-router-dom";
-import GitHubActivityMap from './GitHubActivityMap';
+import GitHubActivityMap from '../GitHubActivityMap/GitHubActivityMap';
 import './Home.css';
 
 const Home = (props) => {
