+++ src/components/Legend/Legend.js
