+++ src/components/ProfileLoader/ProfileLoader.css
@@ -23,7 +23,7 @@
   margin-left: 20px;
   width: 120px;
   min-height: 300px;
-  background-image: url('./sprout.png');
+  background-image: url('../assets/sprout.png');
   background-size: 100%;
   background-repeat: no-repeat;
   background-position: 0 90%;
