+++ src/components/assets/deadplants.png
