+++ src/components/assets/greenhouse-outline.png
