+++ src/components/assets/shoveling_earth.png
