+++ src/components/assets/watering-can.png
